refactor: overhaul data layer architecture and enhance fever tracking
- Refactor all repositories to follow an Interface/Default implementation pattern and move them to the `:core:data:repository` package.
- Introduce `@IoDispatcher` and `@ApplicationScope` qualifiers to centralize coroutine management in the data layer.
- Perform Room database migration to v14:
    - Standardize `childId` types from `String` to `Long` across `MonthlyReport` and related entities.
    - Update `FeverRecord` schema to remove `durationMinutes` and add `isMedicineTaken`.
    - Make `GrowthRecord.aiSuggestion` nullable and standardize primary keys as `Long`.
- Overhaul `:feature:medical` fever tracking:
    - Implement `FeverEpisode` logic to group continuous fever records.
    - Add a Canvas-based `FeverTrendChart` for visualizing temperature fluctuations.
    - Integrate speech recognition (STT) for temperature input and a new tabbed UI layout.
- Update `:core:ui` with `BabySpacing` design tokens and a `LocalBabySpacing` CompositionLocal for standardized layout margins.
- Standardize the "merged/global" data sentinel value as `0L` instead of the previous `String` ("merged") across library and report features.
- Clean up unused models including `WeeklyReport` and consolidate `GrowthSnapshot` definitions.

+++ core/data/src/main/kotlin/com/babymakisuk/coredata/entity/ToiletRecordEntity.kt
@@ -12,7 +12,7 @@ import com.babymakisuk.coremodel.ToiletRecord
     indices = [Index("childId")]
 )
 data class ToiletRecordEntity(
-    @PrimaryKey(autoGenerate = true) val id: Int = 0,
+    @PrimaryKey(autoGenerate = true) val id: Long = 0, // Changed from Int to Long
     val childId: Long,
     val timestamp: Long
 )
