+++ core/model/src/main/kotlin/com/babymakisuk/coremodel/ToiletRecord.kt
@@ -1,7 +1,7 @@
 package com.babymakisuk.coremodel
 
 data class ToiletRecord(
-    val id: Int = 0,
+    val id: Long = 0, // Changed from Int to Long
     val childId: Long,
     val timestamp: Long = System.currentTimeMillis()
 )
