+++ core/model/src/main/kotlin/com/babymakisuk/coremodel/VaccineReminder.kt
@@ -1,7 +1,7 @@
 package com.babymakisuk.coremodel
 
 data class VaccineReminder(
-    val id: Int = 0,
+    val id: Long = 0, // Changed from Int to Long
     val childId: Long,
     val name: String,
     val scheduledDate: Long,
